xlsx TableMetadata impl now calls scratchDB directly.
