+++ drivers/xlsx/ingest.go
@@ -7,6 +7,8 @@ import (
 	"strings"
 	"time"
 
+	"golang.org/x/exp/slices"
+
 	"github.com/neilotoole/sq/libsq/core/options"
 
 	"github.com/neilotoole/sq/libsq/core/lg/lga"
@@ -613,4 +615,27 @@ type sheetTable struct {
 	sheet        *xlsx.Sheet
 	def          *sqlmodel.TableDef
 	hasHeaderRow bool
+}
+
+func detectHeaderRow(sheet *xlsx.Sheet) (hasHeader bool, err error) {
+	if len(sheet.Rows) < 2 {
+		// If zero records, obviously no header row.
+		// If one record... well, is there any way of determining if
+		// it's a header row or not? Probably best to treat it as a data row.
+		return false, nil
+	}
+
+	types1 := getCellColumnTypes(sheet, true)
+	types2 := getCellColumnTypes(sheet, false)
+
+	if len(types1) != len(types2) {
+		// Can this happen?
+		return false, errz.Errorf("sheet {%s} has ragged edges", sheet.Name)
+	}
+
+	if slices.Equal(types1, types2) {
+		return false, nil
+	}
+
+	return true, nil
 }
